scope: unify into 1 file, accumulate into uint16_t
